Fix handling of non-tweened animations
This fixes a misconversion that can happen when working with
frame-by-frame animations -- that is, those drawn directly into the
Timeline, without using Tweens.

The problem was that the converter expects two Shapes that are equal
under object identity comparison (shape1 === shape2) to rasterize to
identical bitmaps.

A fair assumption, but it turns out that animations exported from
Flash Authoring can violate this. It seems Flash keeps the same
Shape object on each frame, but reaches in and messes with its
vector content when the playhead moves.

To fix this issue, we now rasterize Shapes and CRC them to see if
they really are the same or not.

This is mildly expensive, so there's a few early outs before the
actual rasterization happens.